+++ src/transformers/models/jamba/modular_jamba.py
@@ -525,7 +525,7 @@ class JambaSparseMoeBlock(nn.Module):
     def route_tokens_to_experts(self, hidden_states, router_logits):
         routing_weights = torch.nn.functional.softmax(router_logits, dim=-1, dtype=torch.float)
         top_k_weights, top_k_index = torch.topk(routing_weights, self.top_k, dim=-1)
-        return top_k_index, top_k_weights
+        return top_k_index, top_k_weights.to(hidden_states.dtype)
 
     def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
         batch_size, sequence_length, hidden_dim = hidden_states.shape
